Homework 6

+++ Homework_2/app/src/main/java/com/tinkoff/homework/data/DateModel.kt
@@ -1,8 +1,9 @@
 package com.tinkoff.homework.data
 
 import java.time.LocalDate
+import java.util.*
 
 data class DateModel(
-    val id: Int,
+    val id: Long,
     val date: LocalDate,
 )
