+++ Homework_2/app/src/main/java/com/tinkoff/homework/data/dto/StreamDto.kt
@@ -1,9 +1,8 @@
 package com.tinkoff.homework.data.dto
 
 class StreamDto(
-    val id: Int,
+    val id: Long,
     val name: String,
     val topics: List<TopicDto>,
     val isSubscribed: Boolean,
-    var isExpanded: Boolean
-)
+    var isExpanded: Boolean)
