+++ Homework_2/app/src/main/java/com/tinkoff/homework/utils/DelegateItem.kt
@@ -2,6 +2,6 @@ package com.tinkoff.homework.utils
 
 interface DelegateItem {
     fun content(): Any
-    fun id(): Int
+    fun id(): Long
     fun compareToOther(other: DelegateItem): Boolean
 }
