+++ Homework_2/app/src/main/java/com/tinkoff/homework/utils/adapter/date/DateDelegateItem.kt
@@ -4,12 +4,12 @@ import com.tinkoff.homework.data.DateModel
 import com.tinkoff.homework.utils.DelegateItem
 
 class DateDelegateItem(
-    val id: Int,
+    val id: Long,
     private val value: DateModel,
 ) : DelegateItem {
     override fun content(): Any = value
 
-    override fun id(): Int = id
+    override fun id(): Long = id
 
     override fun compareToOther(other: DelegateItem): Boolean {
         return (other as DateDelegateItem).value == content()
