+++ Homework_2/app/src/main/java/com/tinkoff/homework/utils/adapter/stream/StreamDelegateItem.kt
@@ -4,12 +4,12 @@ import com.tinkoff.homework.data.domain.Stream
 import com.tinkoff.homework.utils.DelegateItem
 
 class StreamDelegateItem(
-    val id: Int,
+    val id: Long,
     private val value: Stream,
 ) : DelegateItem {
     override fun content(): Any = value
 
-    override fun id(): Int = id
+    override fun id(): Long = id
 
     override fun compareToOther(other: DelegateItem): Boolean {
         return (other as StreamDelegateItem).value == content()
