+++ Homework_2/app/src/main/java/com/tinkoff/homework/utils/adapter/topic/TopicDelegateItem.kt
@@ -4,12 +4,12 @@ import com.tinkoff.homework.data.domain.Topic
 import com.tinkoff.homework.utils.DelegateItem
 
 class TopicDelegateItem(
-    val id: Int,
+    val id: Long,
     private val value: Topic,
 ) : DelegateItem {
     override fun content(): Any = value
 
-    override fun id(): Int = id
+    override fun id(): Long = id
 
     override fun compareToOther(other: DelegateItem): Boolean {
         return (other as TopicDelegateItem).value == content()
