+++ Homework_2/app/src/main/java/com/tinkoff/homework/view/fragment/ChatFragmentCallback.kt
@@ -3,6 +3,6 @@ package com.tinkoff.homework.view.fragment
 import com.tinkoff.homework.data.Reaction
 
 interface ChatFragmentCallback {
-    fun reactionRemove(reaction: Reaction, messageId: Int)
-    fun showBottomSheetDialog(id: Int): Boolean
+    fun reactionRemove(reaction: Reaction, messageId: Long)
+    fun showBottomSheetDialog(id: Long): Boolean
 }
